adicionando mascara em textfields

+++ lib/functions/drink_helper.dart
@@ -2,7 +2,7 @@ import 'package:agua_diaria/values/dicts.dart';
 import 'package:agua_diaria/values/enums.dart';
 
 double calculateWater(
-    {required int weight,
+    {required double weight,
     required Hour exercisesInMinutes,
     required BiologicSex biologicSex}) {
   return (1.1 * weight + hourDict[exercisesInMinutes]! * 12 / 30) *
